Fix vuepress compile
json-schema-ref-parser doesn't ship with a version compiled for browsers. I
fought for a bit to try and include it in our webpack config, but that didn't
seem to work. Instead, I found an issue https://github.com/APIDevTools/json-schema-ref-parser/issues/192
where someone had turned it into a webpack loader that runs on the schemas at import time.
That seems to have done the trick, and it compiles now.

I also had to change a path for hljs.